优化咩咩的重构
