+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/conditions/segments/MatchSegment.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2011-2020, hubin (jobob@qq.com).
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.baomidou.mybatisplus.core.conditions.segments;
+
+import java.util.function.Predicate;
+
+import com.baomidou.mybatisplus.core.conditions.ISqlSegment;
+import com.baomidou.mybatisplus.core.enums.SqlKeyword;
+
+/**
+ * <p>
+ * 匹配片段
+ * </p>
+ *
+ * @author miemie
+ * @since 2018-06-27
+ */
+public enum MatchSegment {
+    GROUP_BY(i -> i == SqlKeyword.GROUP_BY),
+    ORDER_BY(i -> i == SqlKeyword.ORDER_BY),
+    NOT(i -> i == SqlKeyword.NOT),
+    AND(i -> i == SqlKeyword.AND),
+    OR(i -> i == SqlKeyword.OR),
+    AND_OR(i -> i == SqlKeyword.AND || i == SqlKeyword.OR);
+
+    private final Predicate<ISqlSegment> predicate;
+
+    MatchSegment(Predicate<ISqlSegment> predicate) {
+        this.predicate = predicate;
+    }
+
+    public boolean match(ISqlSegment segment) {
+        return getPredicate().test(segment);
+    }
+
+    protected Predicate<ISqlSegment> getPredicate() {
+        return predicate;
+    }
+}
