+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/conditions/segments/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * SQL 片段相关类
+ */
+package com.baomidou.mybatisplus.core.conditions.segments;
